+++ docs/qgis/edges-cost.qml
@@ -1,0 +1,87 @@
+<!DOCTYPE qgis PUBLIC 'http://mrcc.com/qgis.dtd' 'SYSTEM'>
+<qgis styleCategories="Symbology" version="3.28.0">
+  <renderer-v2 type="graduatedSymbol" attr="cost" graduatedMethod="GraduatedSize" symbollevels="0" forceraster="0" enableorderby="0">
+    <ranges>
+      <range lower="0" upper="200" symbol="0" label="0 - 200" render="true"/>
+      <range lower="200" upper="350" symbol="1" label="200 - 350" render="true"/>
+      <range lower="350" upper="500" symbol="2" label="350 - 500" render="true"/>
+      <range lower="500" upper="700" symbol="3" label="500 - 700" render="true"/>
+      <range lower="700" upper="1000" symbol="4" label="700 - 1000" render="true"/>
+      <range lower="1000" upper="2000" symbol="5" label="1000 - 2000" render="true"/>
+      <range lower="2000" upper="99999" symbol="6" label="2000 - 99999" render="true"/>
+    </ranges>
+    <symbols>
+      <symbol type="line" name="0" alpha="1" clip_to_extent="1" force_rhr="0">
+        <layer class="SimpleLine" enabled="1" locked="0" pass="0">
+          <Option type="Map">
+            <Option type="QString" name="line_color" value="122,31,31,255"/>
+            <Option type="QString" name="line_style" value="solid"/>
+            <Option type="QString" name="line_width" value="0.15"/>
+            <Option type="QString" name="line_width_unit" value="MM"/>
+          </Option>
+        </layer>
+      </symbol>
+      <symbol type="line" name="1" alpha="1" clip_to_extent="1" force_rhr="0">
+        <layer class="SimpleLine" enabled="1" locked="0" pass="0">
+          <Option type="Map">
+            <Option type="QString" name="line_color" value="122,31,31,255"/>
+            <Option type="QString" name="line_style" value="solid"/>
+            <Option type="QString" name="line_width" value="0.3"/>
+            <Option type="QString" name="line_width_unit" value="MM"/>
+          </Option>
+        </layer>
+      </symbol>
+      <symbol type="line" name="2" alpha="1" clip_to_extent="1" force_rhr="0">
+        <layer class="SimpleLine" enabled="1" locked="0" pass="0">
+          <Option type="Map">
+            <Option type="QString" name="line_color" value="122,31,31,255"/>
+            <Option type="QString" name="line_style" value="solid"/>
+            <Option type="QString" name="line_width" value="0.5"/>
+            <Option type="QString" name="line_width_unit" value="MM"/>
+          </Option>
+        </layer>
+      </symbol>
+      <symbol type="line" name="3" alpha="1" clip_to_extent="1" force_rhr="0">
+        <layer class="SimpleLine" enabled="1" locked="0" pass="0">
+          <Option type="Map">
+            <Option type="QString" name="line_color" value="122,31,31,255"/>
+            <Option type="QString" name="line_style" value="solid"/>
+            <Option type="QString" name="line_width" value="0.75"/>
+            <Option type="QString" name="line_width_unit" value="MM"/>
+          </Option>
+        </layer>
+      </symbol>
+      <symbol type="line" name="4" alpha="1" clip_to_extent="1" force_rhr="0">
+        <layer class="SimpleLine" enabled="1" locked="0" pass="0">
+          <Option type="Map">
+            <Option type="QString" name="line_color" value="122,31,31,255"/>
+            <Option type="QString" name="line_style" value="solid"/>
+            <Option type="QString" name="line_width" value="1.1"/>
+            <Option type="QString" name="line_width_unit" value="MM"/>
+          </Option>
+        </layer>
+      </symbol>
+      <symbol type="line" name="5" alpha="1" clip_to_extent="1" force_rhr="0">
+        <layer class="SimpleLine" enabled="1" locked="0" pass="0">
+          <Option type="Map">
+            <Option type="QString" name="line_color" value="122,31,31,255"/>
+            <Option type="QString" name="line_style" value="solid"/>
+            <Option type="QString" name="line_width" value="1.6"/>
+            <Option type="QString" name="line_width_unit" value="MM"/>
+          </Option>
+        </layer>
+      </symbol>
+      <symbol type="line" name="6" alpha="1" clip_to_extent="1" force_rhr="0">
+        <layer class="SimpleLine" enabled="1" locked="0" pass="0">
+          <Option type="Map">
+            <Option type="QString" name="line_color" value="122,31,31,255"/>
+            <Option type="QString" name="line_style" value="solid"/>
+            <Option type="QString" name="line_width" value="2.4"/>
+            <Option type="QString" name="line_width_unit" value="MM"/>
+          </Option>
+        </layer>
+      </symbol>
+    </symbols>
+  </renderer-v2>
+  <layerGeometryType>1</layerGeometryType>
+</qgis>